Put message class inside protocol module
