+++ source/protocol_definitions.py
@@ -10,6 +10,6 @@ CLIENT_PROTOCOL_MAP = protocol.ProtocolMap([
 
 #For communicating with the server
 SERVER_PROTOCOL_MAP = protocol.ProtocolMap([
-    protocol.create_fieldless_message_protocol(BASE_HELP_MESSAGE_PROTOCOL_TYPE_CODE),
+    protocol.create_protocol(BASE_HELP_MESSAGE_PROTOCOL_TYPE_CODE),
     protocol.create_text_message_protocol(HELP_MESSAGE_PROTOCOL_TYPE_CODE)
 ])
